refactor - make consistent types files
