+++ src/components/ActionLink/ActionLink.tsx
@@ -1,52 +1,14 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 import '../../yamui';
 import * as React from 'react';
-import { BaseComponentProps } from '../../util/BaseComponent/props';
 import { join } from '../../util/classNames';
 import Block, { TextSize } from '../Block';
 import Clickable from '../Clickable';
 import { FixedGridRow, FixedGridColumn, GutterSize } from '../FixedGrid';
-import { BaseIcon, IconSize } from '../Icon';
+import { IconSize } from '../Icon';
 import NavigationLink from '../NavigationLink';
 import { getClassNames } from './ActionLink.styles';
-
-export interface BaseActionLinkProps extends BaseComponentProps {
-  /**
-   * The icon component class to be rendered.
-   */
-  icon: typeof BaseIcon;
-
-  /**
-   * A well-curated string describing the action for screen readers.
-   */
-  ariaLabel?: string;
-
-  /**
-   * The visible text.
-   */
-  text: string;
-}
-
-export interface NavigationActionLinkProps extends BaseActionLinkProps {
-  /**
-   * A valid URL.
-   */
-  href: string;
-
-  /**
-   * Whether the link should open in a new window. It securely removes access to the opening window.
-   */
-  newWindow?: boolean;
-}
-
-export interface ClickableActionLinkProps extends BaseActionLinkProps {
-  /**
-   * A click handler.
-   */
-  onClick: ((e: React.MouseEvent<HTMLButtonElement | HTMLLinkElement>) => void);
-}
-
-export type ActionLinkProps = NavigationActionLinkProps | ClickableActionLinkProps;
+import { ActionLinkProps, NavigationActionLinkProps, ClickableActionLinkProps } from './ActionLink.types';
 
 /**
  * An actionable element with icon and text which will either render a button with an onClick callback
