+++ src/components/Clickable/Clickable.tsx
@@ -3,35 +3,8 @@ import '../../yamui';
 import * as React from 'react';
 import FakeLink from '../FakeLink';
 import { join } from '../../util/classNames';
-import { NestableBaseComponentProps, FocusableComponentProps } from '../../util/BaseComponent/props';
 import { getClassNames } from './Clickable.styles';
-export interface ClickableProps extends NestableBaseComponentProps, FocusableComponentProps {
-  /**
-   * Additional label that must be provided if the clickable text is not descriptive enough.
-   */
-  ariaLabel?: string;
-
-  /**
-   * Whether the clickable should be `display: block`.
-   */
-  block?: boolean;
-
-  /**
-   * Title or description of the linked document.
-   */
-  title?: string;
-
-  /**
-   * Whether to remove all styles from the link. Useful for allowing a large area to be clickable
-   * while nesting `FakeLink` components to show link and hover state visuals.
-   */
-  unstyled?: boolean;
-
-  /**
-   * Click callback handler.
-   */
-  onClick?: ((event: React.MouseEvent<HTMLButtonElement>) => void);
-}
+import { ClickableProps } from './Clickable.types';
 
 /**
  * A `Clickable` is an accessible, clickable area that accepts arbitrary children. It is styled
