+++ src/components/Dropdown/Dropdown.tsx
@@ -2,76 +2,14 @@
 import '../../yamui';
 import * as React from 'react';
 import { Dropdown as FabricDropdown, IDropdownOption, DropdownMenuItemType } from 'office-ui-fabric-react/lib/Dropdown';
-import { BaseComponentProps } from '../../util/BaseComponent/props';
 import { join } from '../../util/classNames';
-import { IconSize, BaseIcon } from '../Icon';
+import { IconSize } from '../Icon';
 import ChevronDownMed from '../Icon/icons/ChevronDownMed';
 import Block from '../Block';
 import Text, { TextSize } from '../Text';
 import { FixedGridRow, FixedGridColumn, GutterSize } from '../FixedGrid';
+import { DropdownProps } from './Dropdown.types';
 import './Dropdown.css';
-
-export type DropdownOptionKey = string | number;
-export { DropdownMenuItemType };
-
-export interface DropdownOption {
-  /**
-   * Unique identifier for the given option.
-   */
-  key: DropdownOptionKey;
-
-  /**
-   * The visible option text.
-   */
-  text: string;
-
-  /**
-   * An optional YamUI Icon to be rendered on the left.
-   */
-  icon?: typeof BaseIcon;
-
-  /**
-   * Optional bold text to be displayed before the main text.
-   */
-  label?: string;
-
-  /**
-   * Type of menu item. Specify this for headers or dividers, leave blank for standard options.
-   */
-  itemType?: DropdownMenuItemType;
-}
-
-export interface DropdownProps extends BaseComponentProps {
-  /**
-   * aria-label attribute
-   */
-  ariaLabel?: string;
-
-  /**
-   * Items to be presented to the user.
-   */
-  options: DropdownOption[];
-
-  /**
-   * Label text that will be rendered above the dropdown.
-   */
-  label?: string;
-
-  /**
-   * Input placeholder text. Displayed until option is selected.
-   */
-  placeHolder?: string;
-
-  /**
-   * Initially selected option.
-   */
-  selectedKey?: DropdownOptionKey;
-
-  /**
-   * Function to call when user changes the selected item.
-   */
-  onChanged?: ((key: DropdownOptionKey) => void);
-}
 
 /**
  * A `Dropdown` is a list in which the selected item is always visible, and the others are visible
