+++ src/components/FakeLink/FakeLink.types.ts
@@ -1,3 +1,2 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export { default } from './Clickable';
-export * from './Clickable.types';
+export { NestableBaseComponentProps as FakeLinkProps } from '../../util/BaseComponent/props';
