+++ src/components/FakeLink/index.ts
@@ -1,3 +1,3 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 export { default } from './FakeLink';
-export * from './FakeLink';
+export * from './FakeLink.types';
