+++ src/components/FixedGrid/FixedGridRow.styles.ts
@@ -1,8 +1,8 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 import { mergeStyleSets } from 'office-ui-fabric-react/lib/Styling';
 import { memoizeFunction } from 'office-ui-fabric-react/lib/Utilities';
-import { GutterSize } from './types';
-import { gutterSize as gutterSizes } from '../../util/styles/gutters';
+import { GutterSize } from './FixedGrid.types';
+import { gutterSizes } from '../../util/styles/gutters';
 
 export interface BoxStyleProps {
   bottomSpacing?: GutterSize;
