+++ src/components/FixedGrid/index.ts
@@ -1,5 +1,4 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 export { default as FixedGridRow } from './FixedGridRow';
-export * from './FixedGridRow';
 export { default as FixedGridColumn } from './FixedGridColumn';
-export * from './FixedGridColumn';
+export * from './FixedGrid.types';
