+++ src/components/Heading/Heading.types.ts
@@ -1,0 +1,17 @@
+/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
+import { NestableBaseComponentProps } from '../../util/BaseComponent/props';
+
+export type HeadingLevel = '1' | '2' | '3' | '4' | '5' | '6';
+export type HeadingSize = HeadingLevel | 'none';
+
+export interface HeadingProps extends NestableBaseComponentProps {
+  /**
+   * The heading level to render, i.e. h1-h6
+   */
+  level: HeadingLevel;
+
+  /** Allows a heading tag at a given level to render visually as a different level,
+   * or as plain inline text.
+   */
+  size?: HeadingSize;
+}
