--- src/components/Heading/types.ts
@@ -1,3 +1,0 @@
-/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export type HeadingLevel = '1' | '2' | '3' | '4' | '5' | '6';
-export type HeadingSize = HeadingLevel | 'none';
