+++ src/components/LayoutList/LayoutList.styles.ts
@@ -1,5 +1,5 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-import { gutterSize as gutterSizes } from '../../util/styles/gutters';
+import { gutterSizes } from '../../util/styles/gutters';
 import { mergeStyleSets } from 'office-ui-fabric-react/lib/Styling';
 import { memoizeFunction } from 'office-ui-fabric-react/lib/Utilities';
 import { GutterSize } from '../FixedGrid';
